updates based on PR feedback round 2
