+++ web-ui/src/components/menu/Menu.jsx
@@ -19,11 +19,7 @@ import { AppContext } from '../../context/AppContext';
 import { getAvatarURL } from '../../api/api';
 
 import MenuIcon from '@mui/icons-material/Menu';
-import {
-  useTheme,
-  useColorScheme,
-  Experimental_CssVarsProvider as CssVarsProvider
-} from '@mui/material/styles';
+import { useTheme } from '@mui/material/styles';
 import {
   AppBar,
   Avatar,
@@ -38,7 +34,6 @@ import {
   Modal,
   Toolbar
 } from '@mui/material';
-import { DarkMode, LightMode } from '@mui/icons-material';
 import './Menu.css';
 
 const PREFIX = 'Menu';
@@ -84,7 +79,7 @@ const isCollapsibleListOpen = (linksArr, loc) => {
   return false;
 };
 
-function Menu() {
+function Menu({ children }) {
   const { state, dispatch } = useContext(AppContext);
   const { userProfile } = state;
   const csrf = selectCsrfToken(state);
@@ -265,21 +260,6 @@ function Menu() {
     );
   };
 
-  function SchemeToggle() {
-    const { mode, setMode } = useColorScheme();
-    return (
-      <div
-        className="Menu-modeToggle"
-        onClick={() => {
-          setMode(mode === 'light' ? 'dark' : 'light');
-        }}
-        title={`Select ${mode === 'light' ? 'dark' : 'light'} mode`}
-      >
-        {mode === 'light' ? <DarkMode /> : <LightMode />}
-      </div>
-    );
-  }
-
   const onFileSelected = e => {
     setSelectedFile(e.target.files[0]);
   };
@@ -352,107 +332,105 @@ function Menu() {
             </React.Fragment>
           )}
         </div>
-        <SchemeToggle />
+        {children}
       </List>
     </div>
   );
 
   return (
     <div className={classes.root}>
-      <CssVarsProvider>
-        <CssBaseline />
-        <AppBar position="fixed" className={classes.appBar}>
-          <Toolbar>
-            <IconButton
-              color="inherit"
-              aria-label="open drawer"
-              edge="start"
-              onClick={handleDrawerToggle}
-              className={classes.menuButton}
-              size="large"
-            >
-              <MenuIcon />
-            </IconButton>
-          </Toolbar>
-          <Link to={`/profile/${id}`}>
-            <Avatar
-              src={getAvatarURL(workEmail)}
-              style={{
-                position: 'absolute',
-                cursor: 'pointer',
-                right: '5px',
-                top: '10px',
-                textDecoration: 'none'
-              }}
-              alt={`${userProfile?.name}'s avatar`}
-            />
-          </Link>
-        </AppBar>
-        <nav className={classes.drawer}>
-          <Drawer
-            sx={{ display: { sm: 'none', xs: 'block' } }}
-            variant="temporary"
-            disablePortal
-            anchor={theme.direction === 'rtl' ? 'right' : 'left'}
-            open={mobileOpen}
-            onClose={handleDrawerToggle}
-            classes={{
-              paper: classes.drawerPaper
-            }}
-            ModalProps={{
-              keepMounted: true // Better open performance on mobile.
-            }}
+      <CssBaseline />
+      <AppBar position="fixed" className={classes.appBar}>
+        <Toolbar>
+          <IconButton
+            color="inherit"
+            aria-label="open drawer"
+            edge="start"
+            onClick={handleDrawerToggle}
+            className={classes.menuButton}
+            size="large"
           >
-            {drawer}
-          </Drawer>
-          <Drawer
-            classes={{
-              paper: classes.drawerPaper
+            <MenuIcon />
+          </IconButton>
+        </Toolbar>
+        <Link to={`/profile/${id}`}>
+          <Avatar
+            src={getAvatarURL(workEmail)}
+            style={{
+              position: 'absolute',
+              cursor: 'pointer',
+              right: '5px',
+              top: '10px',
+              textDecoration: 'none'
             }}
-            variant="permanent"
-            open
-            sx={{ display: { xs: 'none', sm: 'block' } }}
-          >
-            {drawer}
-          </Drawer>
-          <Modal
-            open={showHoursUpload}
-            onBackdropClick={closeHoursUpload}
-            onClose={closeHoursUpload}
-          >
-            <div className="hours-upload-modal">
-              <Button color="primary">
-                <label htmlFor="file-upload">
-                  <h3>Choose A CSV File</h3>
-                  <input
-                    accept=".csv"
-                    id="file-upload"
-                    onChange={e => onFileSelected(e)}
-                    style={{ display: 'none' }}
-                    type="file"
-                  />
-                </label>
+            alt={`${userProfile?.name}'s avatar`}
+          />
+        </Link>
+      </AppBar>
+      <nav className={classes.drawer}>
+        <Drawer
+          sx={{ display: { sm: 'none', xs: 'block' } }}
+          variant="temporary"
+          disablePortal
+          anchor={theme.direction === 'rtl' ? 'right' : 'left'}
+          open={mobileOpen}
+          onClose={handleDrawerToggle}
+          classes={{
+            paper: classes.drawerPaper
+          }}
+          ModalProps={{
+            keepMounted: true // Better open performance on mobile.
+          }}
+        >
+          {drawer}
+        </Drawer>
+        <Drawer
+          classes={{
+            paper: classes.drawerPaper
+          }}
+          variant="permanent"
+          open
+          sx={{ display: { xs: 'none', sm: 'block' } }}
+        >
+          {drawer}
+        </Drawer>
+        <Modal
+          open={showHoursUpload}
+          onBackdropClick={closeHoursUpload}
+          onClose={closeHoursUpload}
+        >
+          <div className="hours-upload-modal">
+            <Button color="primary">
+              <label htmlFor="file-upload">
+                <h3>Choose A CSV File</h3>
+                <input
+                  accept=".csv"
+                  id="file-upload"
+                  onChange={e => onFileSelected(e)}
+                  style={{ display: 'none' }}
+                  type="file"
+                />
+              </label>
+            </Button>
+            <div className="buttons">
+              <Button color="secondary" onClick={closeHoursUpload}>
+                Cancel
               </Button>
-              <div className="buttons">
-                <Button color="secondary" onClick={closeHoursUpload}>
-                  Cancel
+              {selectedFile && (
+                <Button
+                  color="primary"
+                  onClick={() => uploadFile(selectedFile)}
+                >
+                  Upload &nbsp;<strong>{selectedFile.name}</strong>
                 </Button>
-                {selectedFile && (
-                  <Button
-                    color="primary"
-                    onClick={() => uploadFile(selectedFile)}
-                  >
-                    Upload &nbsp;<strong>{selectedFile.name}</strong>
-                  </Button>
-                )}
-              </div>
+              )}
             </div>
-          </Modal>
-        </nav>
-        <main className={classes.content}>
-          <div className={classes.toolbar} />
-        </main>
-      </CssVarsProvider>
+          </div>
+        </Modal>
+      </nav>
+      <main className={classes.content}>
+        <div className={classes.toolbar} />
+      </main>
     </div>
   );
 }
